+++ examples/resources/jamfpro_macos_configuration_profile_plist/resource.tf
@@ -56,7 +56,7 @@ resource "jamfpro_macos_configuration_profile_plist" "jamfpro_macos_configuratio
       network_segment_ids                  = [4, 5]
       ibeacon_ids                          = [3, 4]
       directory_service_or_local_usernames = ["Jane Smith", "John Doe"]
-      directory_service_usergroup_ids      = [3, 4]
+      directory_service_usergroup_names    = ["Marketing", "Finance"]
     }
 
     // Optional Block
@@ -69,7 +69,7 @@ resource "jamfpro_macos_configuration_profile_plist" "jamfpro_macos_configuratio
       jss_user_ids                         = [2, 1]
       jss_user_group_ids                   = [4, 505]
       directory_service_or_local_usernames = ["Jane Smith", "John Doe"]
-      directory_service_usergroup_ids      = [3, 4]
+      directory_service_usergroup_names    = ["IT", "Engineering"]
       ibeacon_ids                          = [3, 4]
     }
   }