+++ testing/payloads/jamfpro_computer_extension_attribute/computer_extension_attributes.tf
@@ -10,33 +10,30 @@
 // Single extension attributes
 
 resource "jamfpro_computer_extension_attribute" "jamfpro_computer_extension_attribute_max_script" {
-  name                             = "tf-testing-${var.testing_id}-max-script-${random_id.rng.hex}"
-  description                      = "description"
-  data_type                        = "STRING"
-  enabled                          = true
-  inventory_display_type           = "GENERAL"
-  input_type                       = "SCRIPT"
-  script_contents                  = "echo hello"
-  ldap_extension_attribute_allowed = false
+  name                   = "tf-testing-${var.testing_id}-max-script-${random_id.rng.hex}"
+  description            = "description"
+  data_type              = "STRING"
+  enabled                = true
+  inventory_display_type = "GENERAL"
+  input_type             = "SCRIPT"
+  script_contents        = "echo hello"
 }
 resource "jamfpro_computer_extension_attribute" "jamfpro_computer_extension_attribute_max_text" {
-  name                             = "tf-testing-${var.testing_id}-max-text-${random_id.rng.hex}"
-  description                      = "description"
-  data_type                        = "STRING"
-  enabled                          = true
-  inventory_display_type           = "GENERAL"
-  input_type                       = "TEXT"
-  ldap_extension_attribute_allowed = false
+  name                   = "tf-testing-${var.testing_id}-max-text-${random_id.rng.hex}"
+  description            = "description"
+  data_type              = "STRING"
+  enabled                = true
+  inventory_display_type = "GENERAL"
+  input_type             = "TEXT"
 }
 resource "jamfpro_computer_extension_attribute" "jamfpro_computer_extension_attribute_max_popup" {
-  name                             = "tf-testing-${var.testing_id}-max-popup-${random_id.rng.hex}"
-  description                      = "description"
-  data_type                        = "STRING"
-  enabled                          = true
-  inventory_display_type           = "GENERAL"
-  input_type                       = "POPUP"
-  popup_menu_choices               = ["Option 1", "Option 2", "Option 3"]
-  ldap_extension_attribute_allowed = false
+  name                   = "tf-testing-${var.testing_id}-max-popup-${random_id.rng.hex}"
+  description            = "description"
+  data_type              = "STRING"
+  enabled                = true
+  inventory_display_type = "GENERAL"
+  input_type             = "POPUP"
+  popup_menu_choices     = ["Option 1", "Option 2", "Option 3"]
 }
 
 // Test that minimally defined object is sufficient
@@ -50,34 +47,31 @@ resource "jamfpro_computer_extension_attribute" "jamfpro_computer_extension_attr
 // Multiple extension attributes
 
 resource "jamfpro_computer_extension_attribute" "jamfpro_computer_extension_attribute_max_script_multiple" {
-  count                            = 100
-  name                             = "tf-testing-${var.testing_id}-max-script-${count.index}-${random_id.rng.hex}"
-  description                      = "description"
-  data_type                        = "STRING"
-  enabled                          = true
-  inventory_display_type           = "GENERAL"
-  input_type                       = "SCRIPT"
-  script_contents                  = "echo hello"
-  ldap_extension_attribute_allowed = false
+  count                  = 100
+  name                   = "tf-testing-${var.testing_id}-max-script-${count.index}-${random_id.rng.hex}"
+  description            = "description"
+  data_type              = "STRING"
+  enabled                = true
+  inventory_display_type = "GENERAL"
+  input_type             = "SCRIPT"
+  script_contents        = "echo hello"
 }
 resource "jamfpro_computer_extension_attribute" "jamfpro_computer_extension_attribute_max_text_multiple" {
-  count                            = 100
-  name                             = "tf-testing-${var.testing_id}-max-text-${count.index}-${random_id.rng.hex}"
-  description                      = "description"
-  data_type                        = "STRING"
-  enabled                          = true
-  inventory_display_type           = "GENERAL"
-  input_type                       = "TEXT"
-  ldap_extension_attribute_allowed = false
+  count                  = 100
+  name                   = "tf-testing-${var.testing_id}-max-text-${count.index}-${random_id.rng.hex}"
+  description            = "description"
+  data_type              = "STRING"
+  enabled                = true
+  inventory_display_type = "GENERAL"
+  input_type             = "TEXT"
 }
 resource "jamfpro_computer_extension_attribute" "jamfpro_computer_extension_attribute_max_popup_multiple" {
-  count                            = 100
-  name                             = "tf-testing-${var.testing_id}-max-popup-${count.index}-${random_id.rng.hex}"
-  description                      = "description"
-  data_type                        = "STRING"
-  enabled                          = true
-  inventory_display_type           = "GENERAL"
-  input_type                       = "POPUP"
-  popup_menu_choices               = ["Option 1", "Option 2", "Option 3"]
-  ldap_extension_attribute_allowed = false
+  count                  = 100
+  name                   = "tf-testing-${var.testing_id}-max-popup-${count.index}-${random_id.rng.hex}"
+  description            = "description"
+  data_type              = "STRING"
+  enabled                = true
+  inventory_display_type = "GENERAL"
+  input_type             = "POPUP"
+  popup_menu_choices     = ["Option 1", "Option 2", "Option 3"]
 }